rename to transposeObjectArray
